// packages/diagram/src/likec4diagram/types.d.ts
export declare namespace Types {
    type LeafNodeData = Simplify<BaseNodeData & NonOptional<Pick<DiagramNode, 'id' | 'title' | 'technology' | 'description' | 'notes' | 'color' | 'shape' | 'width' | 'level' | 'height' | 'style' | 'tags' | 'x' | 'y'>> & {
        /**
         * View layout direction, used by DefaultHandles to position node handles
         */
        viewLayoutDir?: AutoLayoutDirection | undefined;
        /**
         * View this node belongs to
         */
        viewId: ViewId;
        isMultiple?: boolean | undefined;
        icon: string | null;
        drifts: NonEmptyReadonlyArray<DiagramNodeDriftReason> | null;
    }>;
    /**
     * Represents element from logical model
     */
    type ElementNodeData = Simplify<LeafNodeData & {
        modelFqn: Fqn;
        deploymentFqn?: never;
        /**
         * If set - this node has navigation to another view and diagram has handler for this
         */
        navigateTo: ViewId | null;
    }>;
    /**
     * Represents element from deployment model
     */
    type DeploymentElementNodeData = Simplify<LeafNodeData & {
        navigateTo: ViewId | null;
        deploymentFqn: DeploymentFqn;
        modelFqn: Fqn | null;
    }>;
// ...
    type CompoundNodeData = Simplify<BaseNodeData & NonOptional<Pick<DiagramNode, 'id' | 'title' | 'notes' | 'color' | 'shape' | 'style' | 'tags' | 'x' | 'y'>> & {
        /**
         * View layout direction, used by DefaultHandles to position node handles
         */
        viewLayoutDir?: AutoLayoutDirection | undefined;
        /**
         * View this node belongs to
         */
        viewId: ViewId;
        depth: number;
        icon?: IconUrl;
        drifts: NonEmptyReadonlyArray<DiagramNodeDriftReason> | null;
    }>;
    type CompoundElementNodeData = Simplify<CompoundNodeData & {
        modelFqn: Fqn;
        deploymentFqn?: never;
        /**
         * If set - this node has navigation to another view and diagram has handler for this
         */
        navigateTo: ViewId | null;
    }>;
    type CompoundDeploymentNodeData = Simplify<CompoundNodeData & {
        deploymentFqn: DeploymentFqn;
        /**
         * If set - this node refers to a model element
         */
        modelFqn: Fqn | null;
        /**
         * If set - this node has navigation to another view and diagram has handler for this
         */
        navigateTo: ViewId | null;
    }>;
    type ViewGroupNodeData = Simplify<CompoundNodeData & {
        isViewGroup: true;
    }>;
    type ElementNode = BaseNode<ElementNodeData, 'element'>;
    type DeploymentElementNode = BaseNode<DeploymentElementNodeData, 'deployment'>;
    type SequenceActorNode = BaseNode<SequenceActorNodeData, 'seq-actor'>;
    type SequenceParallelArea = BaseNode<SequenceParallelAreaData, 'seq-parallel'>;
    type CompoundElementNode = BaseNode<CompoundElementNodeData, 'compound-element'>;
    type CompoundDeploymentNode = BaseNode<CompoundDeploymentNodeData, 'compound-deployment'>;
    type ViewGroupNode = BaseNode<ViewGroupNodeData, 'view-group'>;
    type AnyNode = ElementNode | DeploymentElementNode | CompoundElementNode | CompoundDeploymentNode | ViewGroupNode | SequenceActorNode | SequenceParallelArea;
    type NodeType = AnyNode['type'];
// ...
    type Node<Type extends NodeType = NodeType> = Extract<AnyNode, {
        type: Type;
    }>;
// ...
    type AnyEdge = RelationshipEdge | SequenceStepEdge;
    type EdgeType = AnyEdge['type'];
    type Edge<Type extends EdgeType = EdgeType> = Extract<AnyEdge, {
        type: Type;
    }>;
// ...
}